+++ blocklets/ai-kit/src/pages/playground/playground.tsx
@@ -1,64 +1,73 @@
 import Footer from '@blocklet/ui-react/lib/Footer';
 import Header from '@blocklet/ui-react/lib/Header';
+import styled from '@emotion/styled';
 import { Error, Send } from '@mui/icons-material';
-import { Alert, Box, CircularProgress, IconButton, Input, InputAdornment } from '@mui/material';
+import { Alert, Avatar, Box, BoxProps, CircularProgress, IconButton, Input, InputAdornment } from '@mui/material';
+import { AxiosError } from 'axios';
 import produce from 'immer';
-import { useState } from 'react';
+import { nanoid } from 'nanoid';
+import { ReactNode, useState } from 'react';
 
-import { AIResponse, ai } from '../../libs/ai';
+import { AIResponse, completions } from '../../libs/ai';
 
-let currentId = 0;
-const nextId = () => ++currentId;
+const nextId = () => nanoid(16);
 
 export default function Playground() {
   const [conversations, setConversations] = useState<
-    { id: number; prompt: string; response?: AIResponse; error?: Error }[]
+    { id: string; prompt: string; response?: AIResponse; error?: Error }[]
   >([]);
 
   return (
     <>
-      <Header maxWidth={null} />
+      <Box sx={{ position: 'sticky', zIndex: 100, top: 0 }}>
+        <Header maxWidth={null} />
+      </Box>
 
-      <Box flexGrow={1} sx={{ display: 'flex', flexDirection: 'column', mx: 2 }}>
-        <Box
-          sx={{
-            flex: 1,
-            maxWidth: 800,
-            width: '100%',
-            margin: 'auto',
-            display: 'flex',
-            flexDirection: 'column',
-          }}>
-          <Box flex={1} overflow="auto">
-            {conversations.map((item) => (
-              <Box key={item.id}>
-                <Box my={1}>{item.prompt}</Box>
-                <Box my={1}>
-                  {item.response ? (
-                    <Box whiteSpace="pre-wrap">{item.response?.choices.at(0)?.text}</Box>
-                  ) : item.error ? (
-                    <Alert color="error" icon={<Error />}>
-                      {item.error.message}
-                    </Alert>
-                  ) : (
-                    <CircularProgress size={20} />
-                  )}
-                </Box>
-              </Box>
-            ))}
-          </Box>
-          <Box my={2}>
+      <Box flexGrow={1} m={2}>
+        <Box maxWidth={800} mx="auto" overflow="auto">
+          {conversations.map((item) => (
+            <Box key={item.id} id={`conversation-${item.id}`}>
+              <ConversationItem avatar={<Avatar sx={{ bgcolor: 'secondary.main' }} />}>{item.prompt}</ConversationItem>
+              <ConversationItem
+                my={1}
+                id={`response-${item.id}`}
+                avatar={<Avatar sx={{ bgcolor: 'primary.main' }}>AI</Avatar>}>
+                {item.response ? (
+                  <Box whiteSpace="pre-wrap">{item.response?.choices.at(0)?.text}</Box>
+                ) : item.error ? (
+                  <Alert color="error" icon={<Error />}>
+                    {(item.error as AxiosError<{ message: string }>).response?.data?.message || item.error.message}
+                  </Alert>
+                ) : (
+                  <CircularProgress size={16} />
+                )}
+              </ConversationItem>
+            </Box>
+          ))}
+        </Box>
+      </Box>
+
+      <Box sx={{ position: 'sticky', bottom: 0 }}>
+        <Box height={16} sx={{ pointerEvents: 'none', background: 'linear-gradient(transparent, white)' }} />
+        <Box mx={2} pb={2} sx={{ bgcolor: 'background.paper' }}>
+          <Box maxWidth={800} mx="auto">
             <Prompt
               onSubmit={async (prompt) => {
                 const id = nextId();
                 setConversations((v) => v.concat({ id, prompt }));
+                setTimeout(() => {
+                  document.getElementById(`conversation-${id}`)?.scrollIntoView({ behavior: 'smooth' });
+                });
                 try {
-                  const response = await ai({ prompt });
+                  const response = await completions({ prompt });
                   setConversations((v) =>
                     produce(v, (draft) => {
                       const item = draft.find((i) => i.id === id);
                       if (item) {
-                        item.response = response;
+                        item.response = {
+                          ...response,
+                          choices: response.choices.map((i) => ({ ...i, text: i.text.trim() })),
+                        };
                       }
                     })
                   );
@@ -73,17 +82,41 @@ export default function Playground() {
                   );
 
                   throw error;
+                } finally {
+                  setTimeout(() => {
+                    document.getElementById(`response-${id}`)?.scrollIntoView({ behavior: 'smooth' });
+                  });
                 }
               }}
             />
           </Box>
         </Box>
-      </Box>
 
-      <Footer />
+        <Box sx={{ bgcolor: 'background.paper' }}>
+          <Footer />
+        </Box>
+      </Box>
     </>
   );
 }
+
+function ConversationItem({ children, avatar, ...props }: { children: ReactNode; avatar: ReactNode } & BoxProps) {
+  return (
+    <Box {...props} display="flex">
+      <AvatarWrapper mr={1}>{avatar}</AvatarWrapper>
+      <Box minHeight={30} display="flex" alignItems="center" flex={1} overflow="hidden">
+        {children}
+      </Box>
+    </Box>
+  );
+}
+
+const AvatarWrapper = styled(Box)`
+  > .MuiAvatar-root {
+    width: 30px;
+    height: 30px;
+  }
+`;
 
 function Prompt({ onSubmit }: { onSubmit: (prompt: string) => any }) {
   const [prompt, setPrompt] = useState('');
@@ -93,16 +126,18 @@ function Prompt({ onSubmit }: { onSubmit: (prompt: string) => any }) {
   };
 
   return (
-    <Box sx={{ boxShadow: 2, margin: 'auto', px: 1, borderRadius: 1 }}>
+    <Box
+      component="form"
+      sx={{ boxShadow: 2, margin: 'auto', px: 1, borderRadius: 1 }}
+      onSubmit={(e) => e.preventDefault()}>
       <Input
         fullWidth
         disableUnderline
         value={prompt}
         onChange={(e) => setPrompt(e.target.value)}
-        onKeyDown={(e) => e.key === 'Enter' && submit()}
         endAdornment={
           <InputAdornment position="end">
-            <IconButton onClick={submit} size="small">
+            <IconButton onClick={submit} size="small" type="submit">
               <Send fontSize="small" />
             </IconButton>
           </InputAdornment>