+++ blocklets/ai-kit/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.0.3 (2023-1-30)
+
+- fix: update api path to `/v1/completions`
+- fix: show error message from openai
+- fix: show avatar of conversation
+- fix: auto scroll into view
+- fix: sticky header and footer
+- fix: use nanoid as conversation id
+- fix: use form submit instead of `Enter` listener
+- chore: add bump-version script in workspace root
+- chore: move version file to workspace root
+
 ## 0.0.2 (2023-1-29)
 
 - feat: playground page